+++ src/Ext/Techno/Body.h
@@ -23,6 +23,7 @@ public:
 		ValueableVector<std::unique_ptr<LaserTrailClass>> LaserTrails;
 		Valueable<bool> ReceiveDamage;
 		Valueable<bool> LastKillWasTeamTarget;
+		int PassengerDeletion_Rate;
 
 		ExtData(TechnoClass* OwnerObject) : Extension<TechnoClass>(OwnerObject)
 			, InterceptedBullet(nullptr)
@@ -30,6 +31,7 @@ public:
 			, LaserTrails()
 			, ReceiveDamage(false)
 			, LastKillWasTeamTarget(false)
+			, PassengerDeletion_Rate(-1)
 		{ }
 
 		virtual ~ExtData() = default;
@@ -76,4 +78,5 @@ public:
 	static void ApplyPowered_KillSpawns(TechnoClass* pThis);
 	static void ApplySpawn_LimitRange(TechnoClass* pThis);
 	static void ObjectKilledBy(TechnoClass* pThis, TechnoClass* pKiller);
+	static void EatPassengers(TechnoClass* pThis);
 };