+++ src/Ext/TechnoType/Body.h
@@ -41,6 +41,11 @@ public:
 		Valueable<int> CameoPriority;
 		Valueable<bool> NoManualMove;
 		Nullable<int> InitialStrength;
+		Valueable<bool> PassengerDeletion_Soylent;
+		Valueable<bool> PassengerDeletion_SoylentFriendlies;
+		Valueable<int> PassengerDeletion_Rate;
+		NullableIdx<VocClass> PassengerDeletion_ReportSound;
+		Valueable<bool> PassengerDeletion_Rate_SizeMultiply;
 
 		Valueable<ShieldTypeClass*> ShieldType;
 
@@ -124,6 +129,11 @@ public:
 			LaserTrailData(),
 			DestroyAnim_Random(true),
 			NotHuman_RandomDeathSequence(false),
+			PassengerDeletion_Soylent(false),
+			PassengerDeletion_SoylentFriendlies(false),
+			PassengerDeletion_Rate(0),
+			PassengerDeletion_ReportSound(),
+			PassengerDeletion_Rate_SizeMultiply(true),
 			DefaultDisguise(),
 			OpenTopped_RangeBonus(),
 			OpenTopped_DamageMultiplier(),